requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
docstring_parser==0.17.0
faiss-cpu==1.12.0
fastapi==0.121.2
filetype==1.2.0
frozenlist==1.8.0
google-ai-generativelanguage==0.6.6
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-aiplatform==1.127.0
google-cloud-bigquery==3.38.0
google-cloud-core==2.5.0
google-cloud-resource-manager==1.15.0
google-cloud-storage==3.5.0
google-crc32c==1.7.1
google-genai==1.50.1
google-generativeai==0.7.2
google-resumable-media==2.7.2
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.8
langchain-core==0.3.63
langchain-google-genai==2.0.0
langchain-text-splitters==0.3.8
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.4
langgraph-sdk==0.2.9
langsmith==0.1.147
marshmallow==3.26.1
multidict==6.7.0
mypy_extensions==1.1.0
numexpr==2.14.1
numpy==1.26.4
openai==0.27.8
orjson==3.11.4
ormsgpack==1.12.0
packaging==24.2
pandas==2.3.3
propcache==0.4.1
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.49.3
tenacity==8.5.0
tiktoken==0.12.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
